spike/requirements.txt
kafka-python==1.4.7
django==3.0
psycopg2-binary